Changed namespace from ProductiveRage.Immutable.Extensions to just ProductiveRage.Immutable
I was in two minds about this but there's a definite convenience factor
that the additional classes and extensions become available to existing
code that uses the ProductiveRage.Immutable library, without having to
pull in any more namespaces. The downside is that it could be confusing
if working on a project without this library and wondering why some
classes and methods are not available. I think that it's a good trade
off.